test/requirements.txt
pytest
pyserial
pytest-html